tesseract-macos/scraping_scripts 01-02-47-893/requirements.txt
#install all required python packages by using 'pip install -r requirements.txt'
bs4
requests
pandas
lxml
pymongo
Pillow
opencv-python
numpy
matplotlib
pytesseract